Wwise: complete V136 write support for the interactive music hierarchy
All four interactive-music hirc types now round-trip byte-identically
against the 245 shipped WH3 banks, with independently computed section
sizes: Music_Segment 2093/2093, Music_Track 2798/2798,
Music_Random_Sequence 387/387, Music_Switch 17/17.

HircItem.ReadHirc force-seeks to the declared end of each item after
ReadData returns, so a parser that reads too few bytes is silently
resynced rather than failing. Writing each item back and comparing it to
the bytes it came from is the only thing that actually exercises these
parsers, and it turned up four real defects:

- AkMusicMarkerWwise never read uStringSize before using it to size the
  marker name, so every name came back empty and four bytes were
  swallowed by the resync.
- Several f64 fields (segment duration, marker position, grid period and
  offset, all four clip offsets) were read with ReadInt64 straight into a
  double, reporting values like 4.7e18 instead of 3840.0. ByteParsing has
  no double parser, so the bits are moved across with BitConverter.
- Switch tracks (eTrackType 3) carry a switch association table and a
  transition rule between the track type and iLookAheadTime. These were
  skipped entirely, so LookAheadTime was read out of the middle of the
  switch params for all 63 of them.
- AkPropBundleMinMax wrote each float min/max cast to byte, truncating
  the fraction and wrapping anything negative - a min of -1 was written
  back as 255 and -100 as 156. Fixing this took Sound (379540),
  RandomSequenceContainer (37588) and ActorMixer (3298) from thousands of
  mismatches to none. Same bug fixed in the V112 copy.

The write side of the shared structures the music types sit on was also
missing or wrong: PositioningParams, AuxParams, NodeInitialFxParams,
InitialRtpc and their leaf types had no writer at all, and StateChunk
wrote its two counts in the opposite order to the one it reads them in.

SetState actions were parsed but had neither a writer nor a size branch,
so they could be inspected but never emitted.

Field layouts checked against https://github.com/bnnm/wwiser

+++ Shared/GameFiles/Wwise/Hirc/V112/Shared/AkPropBundleMinMax_V112.cs
@@ -31,14 +31,14 @@ namespace Shared.GameFormats.Wwise.Hirc.V112.Shared
             foreach (var value in PropsList)
                 memStream.Write(ByteParsers.Byte.EncodeValue((byte)value.Type, out _));
 
-            // Read all the Ids first 
+            // These are floats and must not be cast to byte on the way out - doing so truncated
+            // the fraction and wrapped anything negative, so a min of -1 was written back as 255.
             foreach (var value in PropsList)
             {
-                memStream.Write(ByteParsers.Single.EncodeValue((byte)value.Min, out _));
-                memStream.Write(ByteParsers.Single.EncodeValue((byte)value.Max, out _));
+                memStream.Write(ByteParsers.Single.EncodeValue(value.Min, out _));
+                memStream.Write(ByteParsers.Single.EncodeValue(value.Max, out _));
             }
 
-            // Then read the all min and max values.
             var byteArray = memStream.ToArray();
             if (byteArray.Length != GetSize())
                 throw new Exception("Invalid size");
