+++ Shared/GameFiles/Wwise/Hirc/V136/Shared/AkPropBundleMinMax_V136.cs
@@ -29,15 +29,17 @@ namespace Shared.GameFormats.Wwise.Hirc.V136.Shared
             using var memStream = new MemoryStream();
             memStream.Write(ByteParsers.Byte.EncodeValue((byte)PropsList.Count, out _));
 
-            // Read all the Ids first 
+            // First write all the types.
             foreach (var value in PropsList)
                 memStream.Write(ByteParsers.Byte.EncodeValue((byte)value.Type, out _));
 
-            // Then read the all min and max values.
+            // Then write all the min and max values. These are floats and must not be cast to
+            // byte on the way out - doing so truncated the fraction and wrapped anything negative,
+            // so a min of -1 was written back as 255 and -100 as 156.
             foreach (var value in PropsList)
             {
-                memStream.Write(ByteParsers.Single.EncodeValue((byte)value.Min, out _));
-                memStream.Write(ByteParsers.Single.EncodeValue((byte)value.Max, out _));
+                memStream.Write(ByteParsers.Single.EncodeValue(value.Min, out _));
+                memStream.Write(ByteParsers.Single.EncodeValue(value.Max, out _));
             }
 
             return memStream.ToArray();
